requirements.txt
gradio>=4.0,<5.0
starlette>=0.27.0,<0.47
huggingface_hub==0.24.7
llvmlite==0.43.0
numba==0.60.0
tensorflow>=2.13,<3.0
tensorflow-hub>=0.14
tensorflow-datasets>=4.9
librosa>=0.10
numpy>=1.24
Pillow>=10.0
pillow-heif>=0.18
scikit-learn>=1.3
scikit-image>=0.21
joblib>=1.3
pdfplumber>=0.10
PyMuPDF>=1.23
requests>=2.31
beautifulsoup4>=4.12
matplotlib>=3.7
